:memo:  chore: make DraggableAccommodationList component
